Parametrizable WebCrawler
GenericWebCrawler is parametrizable, so it allows GenericCrawlController to parametrize results aswell.
