+++ crawler4j/src/main/java/edu/uci/ics/crawler4j/crawler/GenericCrawlController.java
@@ -46,7 +46,7 @@ import edu.uci.ics.crawler4j.util.IO;
  *
  * @author Yasser Ganjisaffar
  */
-public class GenericCrawlController<CrawlerType extends WebCrawler> {
+public class GenericCrawlController<CrawlerType extends GenericWebCrawler<? extends ResultType>, ResultType> {
 
     static final Logger logger = LoggerFactory.getLogger(GenericCrawlController.class);
     private final CrawlConfig config;
@@ -61,7 +61,7 @@ public class GenericCrawlController<CrawlerType extends WebCrawler> {
      * Once the crawling session finishes the controller collects the local data
      * of the crawler threads and stores them in this List.
      */
-    protected List<Object> crawlersLocalData = new ArrayList<>();
+    protected List<ResultType> crawlersLocalData = new ArrayList<>();
 
     /**
      * Is the crawling of this session finished?
@@ -157,11 +157,11 @@ public class GenericCrawlController<CrawlerType extends WebCrawler> {
         return parser;
     }
 
-    public interface WebCrawlerFactory<T extends WebCrawler> {
+    public interface WebCrawlerFactory<T extends GenericWebCrawler<?>> {
         T newInstance() throws Exception;
     }
 
-    private static class SingleInstanceFactory<T extends WebCrawler>
+    private static class SingleInstanceFactory<T extends GenericWebCrawler<?>>
         implements WebCrawlerFactory<T> {
 
         final T instance;
@@ -176,7 +176,7 @@ public class GenericCrawlController<CrawlerType extends WebCrawler> {
         }
     }
 
-    private static class DefaultWebCrawlerFactory<T extends WebCrawler>
+    private static class DefaultWebCrawlerFactory<T extends GenericWebCrawler<?>>
         implements WebCrawlerFactory<T> {
         final Class<T> clazz;
 
@@ -286,7 +286,7 @@ public class GenericCrawlController<CrawlerType extends WebCrawler> {
                 logger.info("Crawler {} started", i);
             }
 
-            final GenericCrawlController<CrawlerType> controller = this;
+            final GenericCrawlController<CrawlerType, ResultType> controller = this;
             Thread monitorThread = new Thread(new Runnable() {
 
                 @Override
@@ -463,7 +463,7 @@ public class GenericCrawlController<CrawlerType extends WebCrawler> {
      *
      * @return List of Objects which are your local data
      */
-    public List<Object> getCrawlersLocalData() {
+    public List<ResultType> getCrawlersLocalData() {
         return crawlersLocalData;
     }
 
@@ -474,16 +474,16 @@ public class GenericCrawlController<CrawlerType extends WebCrawler> {
             // Do nothing
         }
     }
-    
+
     /**
-     * Extracts data from crawler. Subclases can override this in order to fine-tune 
+     * Extracts data from crawler. Subclases can override this in order to fine-tune
      * results extraction with custom behaviour or processing.
      *
      * @param crawler
      * @return
      */
-    protected Object collectCrawlerData(CrawlerType crawler) {
-    	return crawler.getMyLocalData();
+    protected ResultType collectCrawlerData(CrawlerType crawler) {
+        return crawler.getMyLocalData();
     }
 
     /**
